--- BusinessAnalytics/requirements.txt
@@ -1,5 +1,0 @@
-pandas
-plotly
-openpyxl
-imageio
-kaleido
