+++ requirements.txt
@@ -1,0 +1,5 @@
+pandas
+plotly
+openpyxl
+imageio
+kaleido
